+++ src/types/redis_json.h
@@ -38,7 +38,7 @@ class Json : public Database {
   rocksdb::Status Type(const std::string &user_key, const std::string &path, std::vector<std::string> *results);
   rocksdb::Status ArrAppend(const std::string &user_key, const std::string &path,
                             const std::vector<std::string> &values, std::vector<size_t> *result_count);
-  rocksdb::Status Clear(const std::string &user_key, const std::string &path, int *result);
+  rocksdb::Status Clear(const std::string &user_key, const std::string &path, size_t *result);
 
  private:
   rocksdb::Status write(Slice ns_key, JsonMetadata *metadata, const JsonValue &json_val);
